The act_other.cpp has been fully separated.
