+++ src/engine/network/msdp/msdp_reporters.cpp
@@ -5,6 +5,7 @@
 #include "msdp_constants.h"
 #include "engine/entities/zone.h"
 #include "gameplay/core/game_limits.h"
+#include "gameplay/mechanics/illumination.h"
 
 namespace msdp {
 void RoomReporter::get(Variable::shared_ptr &response) {
