+++ src/engine/ui/alias.h
@@ -1,0 +1,18 @@
+/**
+\file alias.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 19.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_ENGINE_UI_ALIAS_H_
+#define BYLINS_SRC_ENGINE_UI_ALIAS_H_
+
+class CharData;
+void write_aliases(CharData *ch);
+void read_aliases(CharData *ch);
+
+#endif //BYLINS_SRC_ENGINE_UI_ALIAS_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
