+++ src/engine/ui/cmd/do_sneak.cpp
@@ -1,0 +1,47 @@
+/**
+\file do_sneak.cpp - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 19.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#include "engine/entities/char_data.h"
+#include "engine/db/global_objects.h"
+
+void do_sneak(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
+	int prob, percent;
+
+	if (ch->IsNpc() || !ch->GetSkill(ESkill::kSneak)) {
+		SendMsgToChar("Но вы не знаете как.\r\n", ch);
+		return;
+	}
+	if (ch->IsOnHorse()) {
+		act("Вам стоит подумать о мягкой обуви для $N1", false, ch, nullptr, ch->get_horse(), kToChar);
+		return;
+	}
+	if (IsAffectedBySpell(ch, ESpell::kGlitterDust)) {
+		SendMsgToChar("Вы бесшумно крадетесь, отбрасывая тысячи солнечных зайчиков...\r\n", ch);
+		return;
+	}
+	RemoveAffectFromChar(ch, ESpell::kSneak);
+	SendMsgToChar("Хорошо, вы попытаетесь двигаться бесшумно.\r\n", ch);
+	EXTRA_FLAGS(ch).unset(EXTRA_FAILSNEAK);
+	percent = number(1, MUD::Skill(ESkill::kSneak).difficulty);
+	prob = CalcCurrentSkill(ch, ESkill::kSneak, nullptr);
+
+	Affect<EApply> af;
+	af.type = ESpell::kSneak;
+	af.duration = CalcDuration(ch, 0, GetRealLevel(ch), 8, 0, 1);
+	af.modifier = 0;
+	af.location = EApply::kNone;
+	af.battleflag = 0;
+	if (percent > prob) {
+		af.bitvector = 0;
+	} else {
+		af.bitvector = to_underlying(EAffect::kSneak);
+	}
+	affect_to_char(ch, af);
+}
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
