+++ src/engine/ui/cmd/do_visible.h
@@ -1,0 +1,18 @@
+/**
+\file do_visible.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 19.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_ENGINE_UI_CMD_DO_VISIBLE_H_
+#define BYLINS_SRC_ENGINE_UI_CMD_DO_VISIBLE_H_
+
+class CharData;
+void do_visible(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/);
+void perform_immort_vis(CharData *ch);
+
+#endif //BYLINS_SRC_ENGINE_UI_CMD_DO_VISIBLE_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
