+++ src/engine/ui/cmd/do_zone.cpp
@@ -9,6 +9,7 @@
 #include "engine/ui/mapsystem.h"
 #include "gameplay/mechanics/sight.h"
 #include "engine/entities/zone.h"
+#include "gameplay/mechanics/illumination.h"
 
 void DoZone(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
 	if (ch->desc
