+++ src/gameplay/mechanics/groups.h
@@ -17,6 +17,9 @@ void GoGroup(CharData *ch, char *argument);
 void GoUngroup(CharData *ch, char *argument);
 void print_list_group(CharData *ch);
 void print_group(CharData *ch);
+void do_report(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/);
+void do_split(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+void do_split(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/, int currency);
 
 #endif //BYLINS_SRC_GAMEPLAY_MECHANICS_GROUPS_H_
